+++ test/ui/features/chat/views/conversation_task_acceptance_test.dart
@@ -1,0 +1,3 @@
+import '../../../../support/conversation_task_acceptance_flow.dart';
+
+void main() => conversationTaskAcceptanceTests();